Fixed a lot of bugs: - .htaccess improved - Fixed some bugs in DB Maintenance - Fixed few bugs and wrong links on PA File DB - Added ranks images sorting in ACP - Ranks - Fixed few bugs on erc.php - Fixed few bugs on Album - Fixed a small bug on BBCodes and added another couple of features to BBCodes - Added an extra check on STRIP constants on functions.php as well, since those functions are used even when common.php hasn't included - Fixed few bugs on Profile - Fixed a bug in the installation process regarding paths - Fixed a bug when editing a topic with a Poll - Fixed a bug which preventing Shoutbox messages to display email addresses properly - Fixed few bugs on some blocks - Fixed some templates bugs - Fixed a bug when deleting Watched topics - Fixed few minor bugs

+++ watched_topics.php
@@ -43,11 +43,15 @@ if (!$user->data['session_logged_in'])
 // are we un-watching some topics?
 if (isset($_POST['unwatch_list']))
 {
-	$topic_ids = $_POST['unwatch_list'];
-	$sql = "DELETE FROM " . TOPICS_WATCH_TABLE . "
-		WHERE " . $db->sql_in_set('topic_id', $topics_ids) . "
-			AND user_id = " . $user->data['user_id'];
-	$result = $db->sql_query($sql);
+	//$topics_ids = $_POST['unwatch_list'];
+	$topics_ids = request_var('unwatch_list', array(0));
+	if (!empty($topics_ids))
+	{
+		$sql = "DELETE FROM " . TOPICS_WATCH_TABLE . "
+			WHERE " . $db->sql_in_set('topic_id', $topics_ids) . "
+				AND user_id = " . $user->data['user_id'];
+		$result = $db->sql_query($sql);
+	}
 }
 
 include_once(IP_ROOT_PATH . 'includes/users_zebra_block.' . PHP_EXT);
